Adding a workflow to perform submodule synchronization
This leverages the `tut submodule-sync --branch=BRANCH` command and ensures on both `push` and `pull_request` that branches are synchronized. This synchronization will replace our Jenkins job that has been failing for quite some time.

This workflow _does not_ synchronize `master` or `main` branches.

The artifact for this PR is the submodule synchronization check.